Add api key to controller & add some Dao and Models

+++ src/main/java/com/eversoft/traverse/controller/ItineraryListController.java
@@ -17,6 +17,7 @@ import org.springframework.web.bind.annotation.RestController;
 import com.eversoft.traverse.model.User;
 import com.eversoft.traverse.model.Itinerary;
 import com.eversoft.traverse.service.UserService;
+import com.eversoft.traverse.service.AuthService;
 import com.eversoft.traverse.service.ItineraryListService;
 
 @RestController
@@ -26,38 +27,61 @@ public class ItineraryListController {
 	@Autowired
 	ItineraryListService itineraryListService;
 	
+	@Autowired
+	AuthService authService;
+	
 	@RequestMapping(value="/getall", method = RequestMethod.GET, produces=MediaType.APPLICATION_JSON_VALUE)
-	public List<Itinerary> getAllItinerary() {
-		List<Itinerary> list = itineraryListService.getAllItinerary();
-		System.out.println("Itinerary List: " + list);
-		return list;
+	public List<Itinerary> getAllItinerary(@RequestParam(value="key") String key) {
+		if(authService.isValidAPIKey(key)) {
+			List<Itinerary> list = itineraryListService.getAllItinerary();
+			System.out.println("Itinerary List: " + list);
+			return list;
+		}else {
+			return null;
+		}
 	}
 	
 	@RequestMapping(value="/update", method = RequestMethod.POST, consumes=MediaType.APPLICATION_JSON_VALUE)
-	public boolean updateUser(@RequestBody Itinerary itinerary, @RequestParam(value="id") int id) {
-		boolean updated =  itineraryListService.updateItinerary(id, itinerary);
-		System.out.println("Itinerary updated? : " + updated);
-		return updated;
+	public boolean updateUser(@RequestBody Itinerary itinerary, @RequestParam(value="id") int id, @RequestParam(value="key") String key) {
+		if(authService.isValidAPIKey(key)) {
+			boolean updated =  itineraryListService.updateItinerary(id, itinerary);
+			System.out.println("Itinerary updated? : " + updated);
+			return updated;
+		}else {
+			return false;
+		}
 	}
 	
 	@RequestMapping(value="/delete", method = RequestMethod.GET)
-	public boolean deleteUserById(@RequestParam(value="id") int id) {
-		boolean deleted = itineraryListService.deleteItinerary(id);
-		System.out.println("DELETED Itinerary with id = " + id + "?: " + deleted);
-		return deleted;
+	public boolean deleteUserById(@RequestParam(value="id") int id, @RequestParam(value="key") String key) {
+		if(authService.isValidAPIKey(key)) {
+			boolean deleted = itineraryListService.deleteItinerary(id);
+			System.out.println("DELETED Itinerary with id = " + id + "?: " + deleted);
+			return deleted;
+		}else {
+			return false;
+		}
 	}
 	
 	@RequestMapping(value="/add", method = RequestMethod.POST, consumes=MediaType.APPLICATION_JSON_VALUE)
-	public boolean addNewUser(@RequestBody Itinerary itinerary) {
-		boolean added =  itineraryListService.createItinerary(itinerary);
-		System.out.println("NEW Itinerary ADDED? : " + added);
-		return added;
+	public boolean addNewUser(@RequestBody Itinerary itinerary, @RequestParam(value="key") String key) {
+		if(authService.isValidAPIKey(key)) {
+			boolean added =  itineraryListService.createItinerary(itinerary);
+			System.out.println("NEW Itinerary ADDED? : " + added);
+			return added;
+		}else {
+			return false;
+		}
 	}
 	
 	@RequestMapping(value="/get", method = RequestMethod.GET, produces=MediaType.APPLICATION_JSON_VALUE)
-	public Itinerary getItineraryById(@RequestParam(value="id") int id) {
-		Itinerary itinerary = itineraryListService.getItineraryById(id);
-		System.out.println("GET Itinerary: " + itinerary);
-		return itinerary;
+	public Itinerary getItineraryById(@RequestParam(value="id") int id, @RequestParam(value="key") String key) {
+		if(authService.isValidAPIKey(key)) {
+			Itinerary itinerary = itineraryListService.getItineraryById(id);
+			System.out.println("GET Itinerary: " + itinerary);
+			return itinerary;
+		}else {
+			return null;
+		}
 	}
 }
